Add result wrappers to FriendsInvites and groups
Add events for user addition and deletion from group. Changed commands
and commandhandlers for resultwrappers. Add result wrapper behavior for
IActionResult converting in presentation. Changed sync repository
functions to async. Add logs.

TODO: do same to user, problem, auhtentication.

+++ src/Scheduler.Application/FriendsInvites/Commands/AcceptFriendsInvite/AcceptFriendsInviteCommandHandler.cs
@@ -1,5 +1,6 @@
 using MediatR;
 using Scheduler.Application.Common.Interfaces.Persistance;
+using Scheduler.Application.Common.Wrappers;
 using Scheduler.Application.FriendsInvites.Common;
 using Scheduler.Domain.UserAggregate;
 using Scheduler.Domain.UserAggregate.Entities;
@@ -8,24 +9,31 @@ using Scheduler.Domain.UserAggregate.ValueObjects;
 namespace Scheduler.Application.FriendsInvites.Commands.AcceptFriendsInvite;
 
 public class AcceptFriendsInviteCommandHandler(IUsersRepository usersRepository)
-    : IRequestHandler<AcceptFriendsInviteCommand, FriendsInviteResult>
+    : IRequestHandler<AcceptFriendsInviteCommand, ICommandResult<FriendsInviteResult>>
 {
     private readonly IUsersRepository _usersRepository = usersRepository;
 
-    public async Task<FriendsInviteResult> Handle(AcceptFriendsInviteCommand request, CancellationToken cancellationToken)
+    public async Task<ICommandResult<FriendsInviteResult>> Handle(AcceptFriendsInviteCommand request, CancellationToken cancellationToken)
     {
-        User user = await _usersRepository.GetUserByIdAsync(new(request.ExecutorId))
-            ?? throw new NullReferenceException($"No user with id {request.ExecutorId} found.");
-        FriendsInvite friendsInvite = user.ReceivedFriendsInvites.SingleOrDefault(fi => fi.Id.Value == request.InviteId)
-            ?? throw new NullReferenceException($"No friendship invite with id {request.InviteId} found.");
+        User? user = await _usersRepository.GetUserByIdAsync(new(request.ExecutorId));
+        if (user is null)
+        {
+            return new NotFound<FriendsInviteResult>($"No user with id {request.ExecutorId} found.");
+        }
+        FriendsInvite? friendsInvite = user.ReceivedFriendsInvites.SingleOrDefault(fi => fi.Id.Value == request.InviteId);
+        if (friendsInvite is null)
+        {
+            return new NotFound<FriendsInviteResult>($"No friendship invite with id {request.InviteId} found.");
+        }
         user.AcceptFriendshipInvite(new FriendsInviteId(request.InviteId));
         _usersRepository.DeleteFriendsInvite(friendsInvite);
         _usersRepository.Update(user);
         await _usersRepository.SaveChangesAsync();
-        return new FriendsInviteResult(
+        var result = new FriendsInviteResult(
             friendsInvite.Id,
             friendsInvite.SenderId,
             friendsInvite.AddressieId,
             friendsInvite.Message);
+        return new SuccessResult<FriendsInviteResult>(result);
     }
 }
